sistema-reservas/backend/pyproject.toml
[project]
name = "cabanas-reservas-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==4.0.1)",
    "fastapi-mail (>=1.6.1,<2.0.0)",
    "icalendar (>=6.3.2,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "stripe (>=14.3.0,<15.0.0)",
    "cryptography (>=46.0.4,<47.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "alembic (>=1.18.3,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
